refactor: extract message types into room-protocol crate (#197)
Move the Message enum, constructors (make_message, make_join, etc.),
parse_client_line, and all associated tests into a new workspace
member at crates/room-protocol/.

The root crate re-exports room_protocol::* from src/message.rs so
all existing imports (crate::message::*, room::message::*) continue
to work unchanged.

- New crate: crates/room-protocol/ (20 tests)
- Root crate depends on room-protocol via path dependency
- Total test count preserved: 327 (20 + 231 + 71 + 5)

Closes #197
