+++ Cargo.toml
@@ -22,6 +22,7 @@ name = "room"
 path = "src/main.rs"
 
 [dependencies]
+room-protocol = { path = "crates/room-protocol" }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
