+++ crates/room-protocol/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "room-protocol"
+version = "0.1.0"
+edition = "2021"
+description = "Wire format types for the room multi-user chat system"
+license = "MIT"
+repository = "https://github.com/knoxio/room"
+
+[dependencies]
+chrono = { version = "0.4", features = ["serde"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+uuid = { version = "1", features = ["v4"] }
